Allow dialog summary to wrap
Currently even if the summary is extremely long (i.e. select weekly and
check all the days of the week), it just keeps growing in width and it
expands the size of the dialog box as well. I limited the width of the
summary to the width of the dialog.